% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binned_manhattan.R
\name{binned_manhattan_plot.MPdataBinned}
\alias{binned_manhattan_plot.MPdataBinned}
\title{Preprocess GWAS Result for Binned Manhattan Plot}
\usage{
\method{binned_manhattan_plot}{MPdataBinned}(x, ...)
}
\description{
Preprocess GWAS Result for Binned Manhattan Plot
}
