% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binned_manhattan.R
\name{binned_manhattan_plot}
\alias{binned_manhattan_plot}
\title{Preprocess GWAS Result for Binned Manhattan Plot}
\usage{
binned_manhattan_plot(x, ...)
}
\description{
Preprocess GWAS Result for Binned Manhattan Plot
}
