% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds_annotation.R
\name{default_gds_path}
\alias{default_gds_path}
\title{Path to Default GDS File}
\usage{
default_gds_path()
}
\value{
A character vector.
}
\description{
Find path to the default gds file.
}
\examples{
default_gds_path()

}
