% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmanh-package.R
\docType{package}
\name{ggmanh-package}
\alias{ggmanh-package}
\title{ggmanh: Visualization Tool for GWAS Result}
\description{
Manhattan plot and QQ Plot are commonly used to visualize the end result of Genome Wide Association Study. The "ggmanh" package aims to keep the generation of these plots simple while maintaining customizability. Main functions include manhattan_plot, qqunif, and thinPoints.
}
\author{
\strong{Maintainer}: John Lee \email{swannyy.stat@gmail.com}

Authors:
\itemize{
  \item John Lee \email{john.lee@abbvie.com} (AbbVie)
}

Other contributors:
\itemize{
  \item Xiuwen Zheng \email{xiuwen.zheng@abbvie.com} [contributor, data contributor]
}

}
\keyword{internal}
