% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmanh.R
\docType{package}
\name{ggmanh}
\alias{ggmanh}
\title{ggmanh: A package for visualization of GWAS results.}
\description{
ggmanh provides flexible tools for visualizing GWAS result for downstream analysis.
The manhattan plot is generated in two steps: data preprocessing and plotting.
}
