% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{manhattan_chromosome}
\alias{manhattan_chromosome}
\alias{manhattan_chromosome.data.frame}
\alias{manhattan_chromosome.MPdata}
\title{Chromosome Manhattan Plotting}
\usage{
manhattan_chromosome(x, chromosome, ...)

\method{manhattan_chromosome}{data.frame}(
  x,
  chromosome,
  outfn = NULL,
  signif = c(5e-08, 1e-05),
  pval.colname = "pval",
  chr.colname = "chr",
  pos.colname = "pos",
  chr.order = NULL,
  signif.col = NULL,
  chr.col = NULL,
  highlight.colname = NULL,
  preserve.position = FALSE,
  thin = FALSE,
  thin.n = 3000,
  rescale = TRUE,
  rescale.ratio.threshold = 5,
  signif.rel.pos = 0.4,
  label.colname = NULL,
  point.size = 0.75,
  x.label = "Chromosome",
  y.label = expression(-log[10](p)),
  label.font.size = 2,
  max.overlaps = 20,
  plot.title = ggplot2::waiver(),
  plot.subtitle = ggplot2::waiver(),
  plot.width = 10,
  plot.height = 5,
  ...
)

\method{manhattan_chromosome}{MPdata}(
  x,
  chromosome,
  outfn = NULL,
  label.colname = NULL,
  rescale = TRUE,
  rescale.ratio.threshold = 5,
  signif.rel.pos = 0.4,
  color.by.highlight = FALSE,
  point.size = 0.75,
  label.font.size = 2,
  max.overlaps = 20,
  x.label = "Chromosome",
  y.label = expression(-log[10](p)),
  plot.subtitle = ggplot2::waiver(),
  plot.title = ggplot2::waiver(),
  plot.width = 10,
  plot.height = 5,
  ...
)
}
\arguments{
\item{x}{MPdata object.}

\item{chromosome}{a character. A specific chromsome in \code{x} to be plotted}

\item{...}{Additional arguments for chromosome manhattan plot}

\item{outfn}{a character. File name to save the Manhattan Plot. If \code{outfn}
is supplied (i.e. \code{!is.null(outfn)}), then the plot is not drawn in
the graphics window.}

\item{signif}{a numeric vector. Significant p-value thresholds to be drawn for
manhattan plot. At least one value should be provided. Default value is c(5e-08, 1e-5)}

\item{pval.colname}{a character. Column name of \code{x} containing p.value.}

\item{chr.colname}{a character. Column name of \code{x} containing chromosome number.}

\item{pos.colname}{a character. Column name of \code{x} containing position.}

\item{chr.order}{a character vector. Order of chromosomes presented in manhattan plot.}

\item{signif.col}{a character vector of equal length as \code{signif}. It contains
colors for the lines drawn at \code{signif}. If \code{NULL}, the smallest value is colored
black while others are grey.}

\item{chr.col}{a character vector of equal length as chr.order. It contains colors
for the chromosomes. Name of the vector should match \code{chr.order}. If \code{NULL}, default
colors are applied using \code{RColorBrewer}.}

\item{highlight.colname}{a character. If you desire to color certain points
(e.g. significant variants) rather than color by chromosome, you can specify the
category in this column, and provide the color mapping in \code{highlight.col}.
Ignored if \code{NULL}.}

\item{preserve.position}{a logical. If \code{TRUE}, the width of each chromosome reflect the
number of variants and the position of each variant is correctly scaled? If \code{FALSE}, the
width of each chromosome is equal and the variants are equally spaced.}

\item{thin}{a logical. Reduce number of data points when they are cluttered?}

\item{thin.n}{an integer. Number of max points per horizontal partitions of the plot.
Defaults to 500..}

\item{rescale}{a logical. If \code{TRUE}, the plot will rescale itself depending
on the data. More on this in details.}

\item{rescale.ratio.threshold}{a numeric. Threshold of that triggers the rescale.}

\item{signif.rel.pos}{a numeric between 0.1 and 0.9. If the plot is rescaled,
where should the significance threshold be positioned?}

\item{label.colname}{a character. Name of the column in \code{MPdata$data}
to be used for labelling.}

\item{point.size}{a numeric. Size of the points.}

\item{x.label}{a character. x-axis label}

\item{y.label}{a character. y-axis label}

\item{label.font.size}{a numeric. Size of the labels.}

\item{max.overlaps}{an integer. Exclude text labels that overlaps too many things.}

\item{plot.title}{a character. Plot title}

\item{plot.subtitle}{a character. Plot subtitle}

\item{plot.width}{a numeric. Plot width in inches.}

\item{plot.height}{a numeric. Plot height in inches.}

\item{color.by.highlight}{a logical. Should the points be colored based on a highlight column?}
}
\value{
\code{gg} object if \code{is.null(outfn)}, \code{NULL} if \code{!is.null(outf)}

\code{gg} object if \code{is.null(outfn)}, \code{NULL} if \code{!is.null(outf)}
}
\description{
Manhattan Plot of a chosen chromosome
}
\examples{
library(dplyr)

gwasdat <- data.frame(
  "chromosome" = rep(1:5, each = 30),
  "position" = c(replicate(5, sample(1:300, 30))),
  "pvalue" = rbeta(150, 1, 1)^5
)

manhattan_chromosome(
  gwasdat, pval.colname = "pvalue", chromosome = "3",
  chr.colname = "chromosome", pos.colname = "position",
  chr.order = as.character(1:5)
)

library(dplyr)

gwasdat <- data.frame(
  "chromosome" = rep(1:5, each = 30),
  "position" = c(replicate(5, sample(1:300, 30))),
  "pvalue" = rbeta(150, 1, 1)^5
)

mpdata <- manhattan_data_preprocess(
  gwasdat, pval.colname = "pvalue", chr.colname = "chromosome", pos.colname = "position",
  chr.order = as.character(1:5)
)

manhattan_chromosome(mpdata, chromosome = "3")

}
