% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{qqunif}
\alias{qqunif}
\title{Plot Quantile-Quantile Plot of p-values against uniform distribution.}
\usage{
qqunif(x, outfn = NULL, conf.int = 0.95, plot.width = 5, plot.height = 5)
}
\arguments{
\item{x}{a numeric vector of p-values. All values should be between 0 and 1.}

\item{outfn}{a character. File name to save the QQ Plot. If \code{outfn}
is supplied (i.e. \code{!is.null(outfn)}), then the plot is not drawn in
the graphics window.}

\item{conf.int}{a numeric between 0 and 1. Confidence band to draw around reference line. Set to \code{NA} to leave it out.}

\item{plot.width}{a numeric. Plot width in inches.}

\item{plot.height}{a numeric. Plot height in inches.}
}
\value{
a ggplot object
}
\description{
Plot Quantile-Quantile Plot of p-values against uniform distribution.
}
\examples{
x <- rbeta(1000, 1, 1)
qqunif(x)
}
