% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binned_manhattan_preprocess.R
\name{binned_manhattan_preprocess}
\alias{binned_manhattan_preprocess}
\alias{binned_manhattan_preprocess.default}
\alias{binned_manhattan_preprocess.MPdata}
\alias{binned_manhattan_preprocess.data.frame}
\title{Preprocess GWAS Result for Binned Manhattan Plot}
\usage{
binned_manhattan_preprocess(x, ...)

\method{binned_manhattan_preprocess}{default}(x, ...)

\method{binned_manhattan_preprocess}{MPdata}(x, bins_x = 20, bins_y = 100)

\method{binned_manhattan_preprocess}{data.frame}(
  x,
  bins_x = 20,
  bins_y = 100,
  signif = c(5e-08, 1e-05),
  pval.colname = "pval",
  chr.colname = "chr",
  pos.colname = "pos",
  chr.order = NULL,
  signif.col = NULL,
  preserve.position = FALSE,
  pval.log.transform = TRUE,
  ...
)
}
\description{
Preprocess GWAS Result for Binned Manhattan Plot
}
