% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmanh.R
\docType{package}
\name{ggmanh}
\alias{ggmanh}
\title{ggmanh: A package for visualization of GWAS results.}
\description{
ggmanh provides flexible tools for visualizing GWAS result for downstream analysis.
}
\details{
Manhattan plot is commonly used to display significant Single Nucleotide Polymorphisms (SNPs)
in Genome Wide Association Study (GWAS)
This package comes with features useful for manhattan plot creation, including annotation with \code{\link{ggrepel}},
truncating data for faster plot generation, and manual rescaling of the y-axis.
The manhattan plot is generated in two steps: data preprocessing and plotting. This allows the user to iteratively
customize the plot without having the process the GWAS summary data over and over again.
Currently, \code{data.frame} and \code{GRanges} from \code{\link{GenomicRanges}} are supported.

A vignette detailing the usage of the package is accessible by \code{vignette("ggmanh")}
}
